[UV-5R]  "Incorrect 'Model' selected." error
When Baofeng started shipping tri-power radios back in 2014, the firmware
version format could be used to identify if the detected radio had 2 power
levels (HIGH/LOW) or 3 power levels (HIGH/MID/LOW). CHIRP would then use this
firmware version to not only determine if the user had selected the correct
radio model, but also to present the only settings that were available on the
selected model.

Recently Baofeng radio models with 3 power levels have been shipping with with
the HN5RV001 firmware version. This same firmware version, as it has for the
last 2 years, continues to ship in radio models that have only 2 power levels.
This firmware version defeats CHIRP's ability to differentiate on its own
between these two model types.

This patch addresses the problem in 2 ways.

1. for radios with the HN5RV001 firmware version, CHIRP will assume that the
user has selected the correct radio model. The selected model is then appended
to the end of the image so CHIRP can identify the radio model when image is
loaded back into CHIRP at a later time.

2. the driver will now selectively uploads the image in ranges, thus bypassing
specific areas of memory that must not be overwritten by images from radios
with a different firmware version.

This patch also addresses, in a similar way, a related problem where the radio
does not report a firmware version at all to CHIRP.

releated to issue #4445 and many others