[PATCH][tk270.py] Increase OEM band edges by 4.0%, fix #4709
Some of this model's variants are in ranges near ham bands but outside them
by the OEM stated range, for example from 150 to 174 Mhz, or 148 to 174 Mhz,
in the case of the UHF we has the variants in the 450 to 490 Mhz or
the 406 to 430 Mhz.

I make some characterizations and most of the radios can work safely
with a 4% increment of the stated range safely and will allow to work
near ham bands. (real range is bigger but we stick to a safe value)

Related also to #4789, see this for real edges data measurements.
With improvements an corrections suggested on the chirp_devel mailing list